+++ Tests/Integration/Fixtures/can_collect_garbage_from_subPages_when_page_is_set_to_hidden_and_extendToSubpages_is_set.xml
@@ -6,12 +6,14 @@
 		<doktype>1</doktype>
 		<hidden>0</hidden>
 		<extendToSubpages>1</extendToSubpages>
+		<slug>/random</slug>
 	</pages>
 	<pages>
 		<uid>10</uid>
 		<is_siteroot>0</is_siteroot>
 		<doktype>1</doktype>
 		<pid>2</pid>
+		<slug>/random-2</slug>
 	</pages>
 
 	<pages>
@@ -19,5 +21,6 @@
 		<is_siteroot>0</is_siteroot>
 		<doktype>1</doktype>
 		<pid>10</pid>
+		<slug>/random-3</slug>
 	</pages>
 </dataset>
