+++ Tests/Integration/Fixtures/can_find_connection_for_mouted_page.xml
@@ -30,6 +30,7 @@ There is following scenario:
         <clear>3</clear>
         <config>
             <![CDATA[
+                config.index_enable = 1
                 page = PAGE
                 page.typeNum = 0
                 page.bodyTag = <body>
