+++ Tests/Integration/Fixtures/can_queue_a_page_and_remove_it_with_the_garbage_collector.xml
@@ -5,5 +5,6 @@
 		<pid>1</pid>
         <doktype>1</doktype>
         <title>Hello Solr</title>
+        <slug>/</slug>
     </pages>
 </dataset>
