+++ Tests/Integration/Fixtures/can_remove_page.xml
@@ -7,6 +7,7 @@
         <doktype>1</doktype>
         <hidden>0</hidden>
         <title>Test Childpage</title>
+        <slug>/test-childpage</slug>
     </pages>
 
     <tt_content>
