+++ Tests/Integration/Fixtures/does_not_remove_updated_content_element_with_not_set_endtime.xml
@@ -6,6 +6,7 @@
 		<doktype>1</doktype>
 		<hidden>0</hidden>
 		<title>Testpage</title>
+		<slug>/testpage</slug>
 	</pages>
 
 	<tt_content>
